requirements.txt
# AI Image Filter Pipeline - Dependencies

# ============ Core Framework ============
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ============ Data Validation ============
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============ Image Processing ============
Pillow>=10.2.0
imagehash>=4.3.1

# ============ AI/ML ============
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0

# ============ Metadata Analysis ============
# c2pa-python>=0.6.0  # Optional: C2PA support (uncomment if needed)


# ============ Streamlit UI ============
streamlit>=1.31.0
pandas>=2.1.0
requests>=2.31.0

# ============ Development ============
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
